+++ orix/io/plugins/emsoft_h5ebsd.py
@@ -181,7 +181,7 @@ class EMsoftH5ebsdFile(H5ebsdFile):
                 prop = dd[property_name]
                 if prop.shape[-1] == n_top_matches and np.prod(prop.shape) > map_size:
                     # Not a refined dot product file
-                    prop = prop[:map_size].reshape((map_size, n_top_matches))
+                    prop = prop[:map_size].reshape(map_size, n_top_matches)
                 else:
                     # Refined dot product file
                     prop = prop.reshape(map_size)