+++ orix/tests/plot/test_stereographic_plot.py
@@ -235,7 +235,7 @@ class TestStereographicPlot:
     @pytest.mark.parametrize("shape", [(5, 10), (2, 3)])
     def test_multidimensional_vector(self, shape):
         n = np.prod(shape)
-        v = Vector3d(np.random.normal(size=3 * n).reshape(shape + (3,)))
+        v = Vector3d(np.random.normal(size=3 * n).reshape(*shape, 3))
         v.scatter()
         v.draw_circle()