+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Changed
 -------
 - The ``convention`` parameter in ``from_euler()`` and ``to_euler()`` will be removed in
   the next minor release, 0.13, instead of release 1.0 as previously stated.
+- Allow passing a tuple of integers to ``reshape()`` methods of 3D objects.
 
 Deprecated
 ----------